requirements.txt
bitsandbytes==0.43.3
datasets==2.21.0
huggingface_hub==0.24.6
numpy==1.25.2
pandas==2.2.2
python-dotenv==1.0.1
tqdm==4.66.5
transformers==4.44.0
trl==0.9.6
accelerate==0.33.0
peft==0.12.0
scikit-learn
wandb
python-dotenv
